server/requirements.txt
Flask==3.0.0
beautifulsoup4==4.12.2
scrapfly-sdk==0.8.9
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary == 2.9.9